Fixed compilation issues, sql result sets being returned already closed

+++ SkyblockCore-Common/src/main/java/me/illusion/skyblockcore/common/utilities/time/Time.java
@@ -22,7 +22,7 @@ public class Time {
      * @return The converted time
      */
     public int as(TimeUnit unit) {
-        return unit.convert(duration, this.unit);
+        return (int) unit.convert(duration, this.unit);
     }
 
     /**
